fix the lap counter limit

+++ src/network/protocols/server_lobby.cpp
@@ -2072,14 +2072,14 @@ void ServerLobby::startSelection(const Event *event)
                 m_default_vote->m_num_laps =
                     (uint16_t)(UserConfigParams::m_num_goals);
                 if (m_default_vote->m_num_laps > 10)
-                    m_default_vote->m_num_laps = (uint8_t)5;
+                    m_default_vote->m_num_laps = (uint16_t)5;
             }
             else
             {
                 m_default_vote->m_num_laps =
                     (uint8_t)(UserConfigParams::m_soccer_time_limit);
                 if (m_default_vote->m_num_laps > 15)
-                    m_default_vote->m_num_laps = (uint8_t)7;
+                    m_default_vote->m_num_laps = (uint16_t)7;
             }
             m_default_vote->m_reverse = rg.get(2) == 0;
             break;
